# Reader/Writer encoding interfaces