--- a/encoding/marshal_test.go
+++ b/encoding/marshal_test.go
@@ -48,8 +48,8 @@ type (
 	}
 )
 
-func (t test5) MarshalSia() []byte {
-	return AddPrefix([]byte(t.s))
+func (t test5) MarshalSia(w io.Writer) error {
+	return WritePrefix(w, []byte(t.s))
 }
 
 func (t *test5) UnmarshalSia(r io.Reader) error {
@@ -58,9 +58,9 @@ func (t *test5) UnmarshalSia(r io.Reader) error {
 	return err
 }
 
-// reuse above methods, but with a pointer receiver
-func (t *test6) MarshalSia() []byte {
-	return AddPrefix([]byte(t.s))
+// same as above methods, but with a pointer receiver
+func (t *test6) MarshalSia(w io.Writer) error {
+	return WritePrefix(w, []byte(t.s))
 }
 
 func (t *test6) UnmarshalSia(r io.Reader) error {
@@ -107,7 +107,7 @@ func TestEncode(t *testing.T) {
 	for i := range testStructs {
 		err := enc.Encode(testStructs[i])
 		if err != io.ErrShortWrite {
-			t.Error("expected ErrShortWrite, got", err)
+			t.Errorf("testStructs[%d]: expected ErrShortWrite, got %v", i, err)
 		}
 	}
 	// special case, not covered by testStructs