--- a/types/block_bench_test.go
+++ b/types/block_bench_test.go
@@ -8,7 +8,7 @@ import (
 
 // BenchmarkEncodeEmptyBlock benchmarks encoding an empty block.
 //
-// i5-4670K, 09-15-2015: 48 MB/s
+// i5-4670K, 9a90f86: 48 MB/s
 func BenchmarkEncodeBlock(b *testing.B) {
 	var block Block
 	b.SetBytes(int64(len(encoding.Marshal(block))))
@@ -19,8 +19,8 @@ func BenchmarkEncodeBlock(b *testing.B) {
 
 // BenchmarkDecodeEmptyBlock benchmarks decoding an empty block.
 //
-// i7-4770,  08-20-2015: 38 MB/s
-// i5-4670K, 09-15-2015: 55 MB/s
+// i7-4770,  b0b162d: 38 MB/s
+// i5-4670K, 9a90f86: 55 MB/s
 func BenchmarkDecodeEmptyBlock(b *testing.B) {
 	var block Block
 	encodedBlock := encoding.Marshal(block)